# docs/swagger.yaml
definitions:
  vaultex_internal_shared.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  vaultex_pkg_dto.CreateOrganizationRequest:
    properties:
      email:
        type: string
      organization_name:
        type: string
      password:
        minLength: 8
        type: string
      phone_number:
        type: string
      registered_name:
        type: string
      website_url:
        type: string
    required:
    - email
    - organization_name
    - password
    - phone_number
    - registered_name
    type: object
  vaultex_pkg_dto.CreateWalletRequest:
    properties:
      currency:
        type: string
      external_user_email:
        type: string
      external_user_id:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      organization_id:
        type: string
    required:
    - currency
    - external_user_id
    - organization_id
    type: object
  vaultex_pkg_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  vaultex_pkg_dto.OrganizationLoginResponse:
    properties:
      api_key:
        type: string
      email:
        type: string
      id:
        type: string
      organization_name:
        type: string
      phone_number:
        type: string
      registered_name:
        type: string
    type: object
  vaultex_pkg_dto.OrganizationSignupResponse:
    properties:
      api_key:
        type: string
      api_secret:
        type: string
      email:
        type: string
      id:
        type: string
      organization_name:
        type: string
      phone_number:
        type: string
      registered_name:
        type: string
    type: object
info:
  contact:
    email: osundemichael7@gmail.com
    name: API Support
    url: https://github.com/michaelkinginfosec
  description: A Ledger-as-a-Service API. Platforms integrate Vaultex to manage their
    users' wallets and track every financial transaction via double-entry bookkeeping.
  title: Vaultex
  version: "1.0"
paths:
  /organizations:
    post:
      consumes:
      - application/json
      description: Create a new organization with the provided name and email
      parameters:
      - description: Organization data
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/vaultex_pkg_dto.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/vaultex_pkg_dto.OrganizationSignupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
      summary: Create a new organization
      tags:
      - Auth
  /organizations/email/{email}:
    get:
      consumes:
      - application/json
      description: Retrieve an organization by their email address
      parameters:
      - description: Organization Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vaultex_pkg_dto.OrganizationLoginResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
      summary: Get organization by email
      tags:
      - organizations
  /organizations/login:
    post:
      consumes:
      - application/json
      description: Authenticate an organization using their email and password
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/vaultex_pkg_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vaultex_pkg_dto.OrganizationLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
      summary: Login an organization
      tags:
      - Auth
  /wallets:
    post:
      consumes:
      - application/json
      description: Create a wallet for an external user under the authenticated organization
      parameters:
      - description: API Key
        in: header
        name: x-api-key
        required: true
        type: string
      - description: HMAC Signature
        in: header
        name: x-signature
        required: true
        type: string
      - description: Unix Timestamp
        in: header
        name: x-timestamp
        required: true
        type: string
      - description: Create wallet request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/vaultex_pkg_dto.CreateWalletRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Create user wallet
      tags:
      - Wallets
  /wallets/{external_user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a wallet using the authenticated organization and external
        user ID
      parameters:
      - description: API Key
        in: header
        name: x-api-key
        required: true
        type: string
      - description: HMAC Signature
        in: header
        name: x-signature
        required: true
        type: string
      - description: Unix Timestamp
        in: header
        name: x-timestamp
        required: true
        type: string
      - description: External User ID
        in: path
        name: external_user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vaultex_internal_shared.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get wallet by external user ID
      tags:
      - Wallets
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"